requirements.txt
google-auth-oauthlib==1.2.0
google-api-python-client==2.137.0
python-telegram-bot==21.4
python-dotenv==1.0.1
requests==2.32.3